+++ apps/decodex/src/orchestrator/tests/runtime_loop_scenarios/loop_scenarios/support.rs
@@ -1,0 +1,7 @@
+mod fixtures;
+mod harness;
+
+pub(super) use self::{
+	fixtures::loop_scenario_repo_gate_failure,
+	harness::{LOOP_SCENARIO_GATE_SERVICE_ID, LoopScenarioHarness},
+};
